refactor(character): csettings cli
