+++ cogs5e/models/character.py
@@ -191,7 +191,11 @@ class Character(StatBlock):
 
     # ---------- Basic CRUD ----------
     def get_color(self) -> int:
-        return self.options.get('color') or super().get_color()
+        color_option = self.options.color
+        if color_option is None:
+            return super().get_color()
+        r, g, b = color_option.as_rgb_tuple(alpha=False)
+        return (r << 16) + (g << 8) + b
 
     @property
     def owner(self) -> str:
