+++ ui/__init__.py
@@ -1,1 +1,2 @@
 from .servsettings import ServerSettingsUI
+from .charsettings import CharacterSettingsUI
